Removed BufferView, clippy should be happy now

+++ crates/common/src/memory.rs
@@ -282,14 +282,14 @@ impl BlockAllocator {
 
     pub fn allocate(&mut self) -> Option<u32> {
         if let Some(slot) = self.empty.pop() {
-            Some((slot as u32) * self.chunk_size)
+            Some(slot * self.chunk_size)
         } else {
             None
         }
     }
 
     pub fn dealloc(&mut self, offset: u32) {
-        let index = (offset / self.chunk_size) as u32;
+        let index = offset / self.chunk_size;
         assert!(index < self.chunk_count && offset % self.chunk_size == 0);
         assert!(!self.empty.contains(&index));
         self.empty.push(index);
